tắt câu query

+++ admin/views/controller-nguyen-lieu.php
@@ -36,7 +36,7 @@ class NguyenLieu extends xuly
         $ten_nguyen_lieu = $data_insert->ten_nguyen_lieu;
         if(empty($data_insert->ten_nguyen_lieu)) return 'ten_nguyen_lieu';
 
-        $so_luong = (int)$data_insert->so_luong;
+        $so_luong = (float)$data_insert->so_luong;
         if($so_luong < 0) return 'so_luong';
 
         if(empty($data_insert->gia_tien)) return 'gia_tien';
@@ -73,7 +73,7 @@ class NguyenLieu extends xuly
         $ten_nguyen_lieu = $data_update->ten_nguyen_lieu;
         if(empty($data_update->ten_nguyen_lieu)) return 'ten_nguyen_lieu';
 
-        $so_luong = (int)$data_update->so_luong;
+        $so_luong = (float)$data_update->so_luong;
         if($so_luong < 0) return 'so_luong';
 
         if(empty($data_update->gia_tien)) return 'gia_tien';
